Move local_cs tests to own file
